refactor(mllp-client,mllp-transport): streams-native exchange pipeline
Replaces the bespoke `exchange()` orchestrator in mllp-client with a
`socket → decoder → ack-parser` pipeThrough pipeline. The decoder now
runs in fatal mode (its onError throws), the new ack-parser
TransformStream parses each decoded frame into a typed Acknowledgment
or errors the readable on a malformed ACK. MllpClient.send() consumes
parsed acknowledgments directly from the reader, which collapses the
manual reader.read() loop, the inline parseAck call, the localAbort
controller used to surface decoder errors, and the bespoke lock
cleanup.

mllp-transport: formalise the lenient/fatal contract on
DecoderOptions.onError. Throwing from onError propagates out of
transform() and errors the readable side of the TransformStream;
returning normally keeps the decoder scanning for the next valid
frame. Three new tests pin the contract (garbage bytes, oversized
frame, custom error class preservation). README and JSDoc updated to
document both modes with examples.

mllp-client: delete core/internal/exchange.ts (322 lines). Inline the
orchestration into client.ts as a private readAck() function plus
small focused helpers. Add core/internal/ack-parser-stream.ts (5-line
transform). All 42 tests pass without modification. Bundle size drops
from 23.02 to 22.45 kB minified.

A separate GitHub issue tracks migrating @glion/mllp (server) to the
same streams-native pattern.
