+++ packages/mllp-client/src/runtimes/deno.ts
@@ -142,10 +142,10 @@ export const denoConnect: MllpConnect = async (params) => {
 };
 
 /**
- * Translate an abort `reason` from the connect-phase signal into a
- * typed {@link MllpClientError}. Mirrors the precedence in
- * `normaliseExchangeError`: typed reasons pass through, anything
- * else maps to TIMEOUT with the reason chained as `cause`.
+ * Translate an abort `reason` from the connect-phase signal into a typed
+ * {@link MllpClientError}. Mirrors the precedence in `normaliseSendError` in
+ * `core/client.ts`: typed reasons pass through, anything else maps to TIMEOUT
+ * with the reason chained as `cause`.
  */
 function toAbortError(
   reason: unknown,
