+++ packages/mllp-transport/src/types.ts
@@ -33,7 +33,24 @@ export interface DecoderOptions {
   maxMessageSize?: number;
   /** Encoding for text conversion (default: 'utf-8') */
   encoding?: string;
-  /** Error callback for logging/handling errors */
+  /**
+   * Called for each frame-level error (invalid start byte, oversized
+   * frame, incomplete frame at stream end). Two modes are supported,
+   * selected by what the callback does:
+   *
+   * - **Lenient (return normally)** — the decoder logs/handles the error and
+   *   keeps scanning the buffer for the next valid frame. This is the right
+   *   default for receivers that process many frames over a long-lived
+   *   connection: a single bad frame should not tear down the stream.
+   * - **Fatal (throw)** — throwing from inside the callback propagates out of the
+   *   underlying `transform()` and errors the readable side of the
+   *   `TransformStream`. Downstream `for await ... of` consumers reject with
+   *   the thrown value. This is the right mode for one-shot exchanges (e.g. the
+   *   MLLP client) where any frame error invalidates the whole exchange.
+   *
+   * The default callback (when this option is omitted) logs to
+   * `console.warn` and returns — i.e. lenient.
+   */
   onError?: (error: TransportError) => void;
 }
